Add client side data
This commit adds Watermelon DB as a persistence layer for the application. It
also wires up the functionality to add a video and remember it for later.

I think, long term, these videos should be unique so it's easier to track
progress. Also I think I need to investigate the observable layer to make sure
data is refreshed when returning home.